Add files via upload
This Golang project receives as inputs text files describing an ants' colony, e.g. the number of ants, the number of rooms, and the tunnels that connect the rooms. The objective is to build a graph based on this information, determine the most efficient path(s) from the 'Start' room to the 'End' room, and marshal the ants through these paths in the quickest way.